requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.6
django-rest-swagger==2.1.2
djangorestframework==3.7.1
djangorestframework-jwt==1.11.0
factory-boy==2.9.2
Faker==0.8.6
flake8==3.5.0
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
mccabe==0.6.1
openapi-codec==1.3.2
py==1.4.34
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.5.3
pytest==3.2.3
pytest-django==3.1.2
pytest-sugar==0.9.0
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
simplejson==3.11.1
six==1.11.0
termcolor==1.1.0
text-unidecode==1.0
uritemplate==3.0.0
urllib3==1.22
